requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
cryptography==43.0.1
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
lxml==5.3.0
rich==13.9.2
pytest==8.3.3
